Update backend to key off name instead of ID
So that you can go on your phone or computer and see what your response is.

There's no protections - if you want to pretend to be someone else and change their responses, you can.
This also means that if you happen to have the same name as someone else, you should see that pretty quickly.